+++ CHANGELOG.md
@@ -28,6 +28,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [UNRELEASED]
 
+## v0.1.7 - 2024-08-27
+- Added CF worker namespace listing and selection.
+- Removed manual namespaceId input in settings dialog.
+
 ## v0.1.6 - 2024-08-27
 - Added option to manually check for app updates.
 - Updated Tauri dependencies.